+++ crates/networking/p2p/rlpx/error.rs
@@ -51,6 +51,8 @@ pub(crate) enum RLPxError {
     IoError(#[from] std::io::Error),
     #[error("Failed to decode message due to invalid frame: {0}")]
     InvalidMessageFrame(String),
+    #[error("IncompatibleProtocol")]
+    IncompatibleProtocol,
 }
 
 // tokio::sync::mpsc::error::SendError<Message> is too large to be part of the RLPxError enum directly
