+++ crates/networking/p2p/rlpx/eth/backend.rs
@@ -1,46 +1,14 @@
-use ethrex_common::{types::ForkId, U256};
+use ethrex_common::types::ForkId;
 use ethrex_storage::Store;
 
-use crate::rlpx::error::RLPxError;
+use crate::rlpx::{error::RLPxError, p2p::Capability};
 
 use super::status::StatusMessage;
-
-pub async fn get_status(storage: &Store, eth_version: u32) -> Result<StatusMessage, RLPxError> {
-    let chain_config = storage.get_chain_config()?;
-    let total_difficulty = U256::from(chain_config.terminal_total_difficulty.unwrap_or_default());
-    let network_id = chain_config.chain_id;
-
-    // These blocks must always be available
-    let genesis_header = storage
-        .get_block_header(0)?
-        .ok_or(RLPxError::NotFound("Genesis Block".to_string()))?;
-    let block_number = storage.get_latest_block_number().await?;
-    let block_header = storage
-        .get_block_header(block_number)?
-        .ok_or(RLPxError::NotFound(format!("Block {block_number}")))?;
-
-    let genesis = genesis_header.compute_block_hash();
-    let block_hash = block_header.compute_block_hash();
-    let fork_id = ForkId::new(
-        chain_config,
-        genesis_header,
-        block_header.timestamp,
-        block_number,
-    );
-    Ok(StatusMessage {
-        eth_version,
-        network_id,
-        total_difficulty,
-        block_hash,
-        genesis,
-        fork_id,
-    })
-}
 
 pub async fn validate_status(
     msg_data: StatusMessage,
     storage: &Store,
-    eth_version: u32,
+    eth_capability: &Capability,
 ) -> Result<(), RLPxError> {
     let chain_config = storage.get_chain_config()?;
 
@@ -61,26 +29,26 @@ pub async fn validate_status(
     );
 
     //Check networkID
-    if msg_data.network_id != chain_config.chain_id {
+    if msg_data.get_network_id() != chain_config.chain_id {
         return Err(RLPxError::HandshakeError(
             "Network Id does not match".to_string(),
         ));
     }
     //Check Protocol Version
-    if msg_data.eth_version != eth_version {
+    if msg_data.get_eth_version() != eth_capability.version {
         return Err(RLPxError::HandshakeError(
             "Eth protocol version does not match".to_string(),
         ));
     }
     //Check Genesis
-    if msg_data.genesis != genesis_hash {
+    if msg_data.get_genesis() != genesis_hash {
         return Err(RLPxError::HandshakeError(
             "Genesis does not match".to_string(),
         ));
     }
     // Check ForkID
     if !fork_id.is_valid(
-        msg_data.fork_id,
+        msg_data.get_fork_id(),
         latest_block_number,
         latest_block_header.timestamp,
         chain_config,
@@ -95,11 +63,14 @@ pub async fn validate_status(
 #[cfg(test)]
 mod tests {
     use super::validate_status;
+    use crate::rlpx::eth::eth68::status::StatusMessage68;
     use crate::rlpx::eth::status::StatusMessage;
+    use crate::rlpx::p2p::Capability;
     use ethrex_common::{
         types::{ForkId, Genesis},
         H256, U256,
     };
+
     use ethrex_storage::{EngineType, Store};
     use std::{fs::File, io::BufReader};
 
@@ -125,16 +96,16 @@ mod tests {
         let genesis_hash = genesis_header.compute_block_hash();
         let fork_id = ForkId::new(config, genesis_header, 2707305664, 123);
 
-        let eth_version = 68;
-        let message = StatusMessage {
-            eth_version,
+        let eth = Capability::eth(68);
+        let message = StatusMessage::StatusMessage68(StatusMessage68 {
+            eth_version: eth.version,
             network_id: 3503995874084926,
             total_difficulty,
             block_hash: H256::random(),
             genesis: genesis_hash,
             fork_id,
-        };
-        let result = validate_status(message, &storage, eth_version).await;
+        });
+        let result = validate_status(message, &storage, &eth).await;
         assert!(result.is_ok());
     }
 }
