+++ crates/networking/p2p/rlpx/eth/mod.rs
@@ -1,5 +1,6 @@
 pub(crate) mod backend;
 pub(crate) mod blocks;
+mod eth68;
 pub(crate) mod receipts;
 pub(crate) mod status;
 pub(crate) mod transactions;
