+++ crates/networking/p2p/rlpx/eth/status.rs
@@ -1,67 +1,100 @@
-use crate::rlpx::{
-    message::RLPxMessage,
-    utils::{snappy_compress, snappy_decompress},
-};
+use super::eth68::status::StatusMessage68;
+use crate::rlpx::message::RLPxMessage;
+use crate::rlpx::utils::snappy_decompress;
+use crate::rlpx::{error::RLPxError, p2p::Capability};
 use bytes::BufMut;
-use ethrex_common::{
-    types::{BlockHash, ForkId},
-    U256,
-};
-use ethrex_rlp::{
-    error::{RLPDecodeError, RLPEncodeError},
-    structs::{Decoder, Encoder},
-};
+use ethrex_common::types::{BlockHash, ForkId};
+use ethrex_common::U256;
+use ethrex_rlp::error::{RLPDecodeError, RLPEncodeError};
+use ethrex_rlp::structs::Decoder;
+use ethrex_storage::Store;
 
 #[derive(Debug)]
-pub(crate) struct StatusMessage {
-    pub(crate) eth_version: u32,
-    pub(crate) network_id: u64,
-    pub(crate) total_difficulty: U256,
-    pub(crate) block_hash: BlockHash,
-    pub(crate) genesis: BlockHash,
-    pub(crate) fork_id: ForkId,
+pub enum StatusMessage {
+    StatusMessage68(StatusMessage68),
 }
 
 impl RLPxMessage for StatusMessage {
     const CODE: u8 = 0x10;
     fn encode(&self, buf: &mut dyn BufMut) -> Result<(), RLPEncodeError> {
-        let mut encoded_data = vec![];
-        Encoder::new(&mut encoded_data)
-            .encode_field(&self.eth_version)
-            .encode_field(&self.network_id)
-            .encode_field(&self.total_difficulty)
-            .encode_field(&self.block_hash)
-            .encode_field(&self.genesis)
-            .encode_field(&self.fork_id)
-            .finish();
-
-        let msg_data = snappy_compress(encoded_data)?;
-        buf.put_slice(&msg_data);
-        Ok(())
+        match self {
+            StatusMessage::StatusMessage68(msg) => msg.encode(buf),
+        }
     }
 
     fn decode(msg_data: &[u8]) -> Result<Self, RLPDecodeError> {
         let decompressed_data = snappy_decompress(msg_data)?;
         let decoder = Decoder::new(&decompressed_data)?;
-        let (eth_version, decoder): (u32, _) = decoder.decode_field("protocolVersion")?;
+        let (eth_version, _): (u32, _) = decoder.decode_field("protocolVersion")?;
 
-        assert_eq!(eth_version, 68, "only eth version 68 is supported");
+        match eth_version {
+            68 => Ok(StatusMessage::StatusMessage68(StatusMessage68::decode(
+                msg_data,
+            )?)),
+            _ => Err(RLPDecodeError::IncompatibleProtocol),
+        }
+    }
+}
 
-        let (network_id, decoder): (u64, _) = decoder.decode_field("networkId")?;
-        let (total_difficulty, decoder): (U256, _) = decoder.decode_field("totalDifficulty")?;
-        let (block_hash, decoder): (BlockHash, _) = decoder.decode_field("blockHash")?;
-        let (genesis, decoder): (BlockHash, _) = decoder.decode_field("genesis")?;
-        let (fork_id, decoder): (ForkId, _) = decoder.decode_field("forkId")?;
-        // Implementations must ignore any additional list elements
-        let _padding = decoder.finish_unchecked();
+impl StatusMessage {
+    pub async fn new(storage: &Store, eth: &Capability) -> Result<Self, RLPxError> {
+        let chain_config = storage.get_chain_config()?;
+        let total_difficulty =
+            U256::from(chain_config.terminal_total_difficulty.unwrap_or_default());
+        let network_id = chain_config.chain_id;
 
-        Ok(Self {
-            eth_version,
-            network_id,
-            total_difficulty,
-            block_hash,
-            genesis,
-            fork_id,
-        })
+        // These blocks must always be available
+        let genesis_header = storage
+            .get_block_header(0)?
+            .ok_or(RLPxError::NotFound("Genesis Block".to_string()))?;
+        let block_number = storage.get_latest_block_number().await?;
+        let block_header = storage
+            .get_block_header(block_number)?
+            .ok_or(RLPxError::NotFound(format!("Block {block_number}")))?;
+
+        let genesis = genesis_header.compute_block_hash();
+        let block_hash = block_header.compute_block_hash();
+        let fork_id = ForkId::new(
+            chain_config,
+            genesis_header,
+            block_header.timestamp,
+            block_number,
+        );
+
+        match eth.version {
+            68 => Ok(StatusMessage::StatusMessage68(StatusMessage68 {
+                eth_version: eth.version,
+                network_id,
+                total_difficulty,
+                block_hash,
+                genesis,
+                fork_id,
+            })),
+            _ => Err(RLPxError::IncompatibleProtocol),
+        }
+    }
+
+    pub fn get_network_id(&self) -> u64 {
+        match self {
+            StatusMessage::StatusMessage68(msg) => msg.network_id,
+        }
+    }
+
+    pub fn get_eth_version(&self) -> u8 {
+        match self {
+            StatusMessage::StatusMessage68(msg) => msg.eth_version,
+        }
+    }
+
+    pub fn get_fork_id(&self) -> ForkId {
+        match self {
+            StatusMessage::StatusMessage68(msg) => msg.fork_id.clone(),
+        }
+    }
+
+    pub fn get_genesis(&self) -> BlockHash {
+        match self {
+            StatusMessage::StatusMessage68(msg) => msg.genesis,
+        }
     }
 }
